Extract search api response stubs from the existing OrganisationsApiController tests - create new helper module OrganisationsApiTestHelper - these can be then be reused for the pact tests in subsequent commits.
